Fixed bug computing norm_estimate

+++ hydra/algorithms/time_evolution/time_evolution.h
@@ -9,12 +9,12 @@ namespace hydra {
 std::tuple<double, double> time_evolve_inplace(BondList const &bonds,
                                                State &state, double time,
                                                double precision = 1e-12,
-                                               int m = 5, double anorm = 0.,
-                                               int nnorm = 2);
+                                               int64_t m = 5, double anorm = 0.,
+                                               int64_t nnorm = 2);
 
 State time_evolve(BondList const &bonds, State state, double time,
-                  double precision = 1e-12, int m = 5, double anorm = 0.,
-                  int nnorm = 2);
+                  double precision = 1e-12, int64_t m = 5, double anorm = 0.,
+                  int64_t nnorm = 2);
 
 double imag_time_evolve_inplace(BondList const &bonds, State &state,
                                 double time, double precision = 1e-12);
